Authorizing Users

+++ app/controllers/access_tokens_controller.rb
@@ -2,6 +2,7 @@ class AccessTokensController < ApplicationController
   before_action :authenticate_user, only: :destroy
 
   def create
+    skip_authorization
     user = User.find_by!(email: login_params[:email])
 
     if user.authenticate(login_params[:password])
@@ -24,6 +25,7 @@ class AccessTokensController < ApplicationController
   end
 
   def destroy
+    authorize access_token
     access_token.destroy
     render status: :no_content
   end
